+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Retry with exponential backoff function to error handling
+
 ## 2.2.13 - 2021-09-21
 
 ### Fixed